+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Dev: Save settings in `aboutToQuit`. (#6159)
 - Dev: Bumped deprecation cutoff to Qt 6.4.3. (#6169)
 - Dev: Updated GoogleTest to v1.17.0. (#6180)
+- Dev: Mini refactor of `TwitchAccount`. (#6182)
 - Dev: Simplified string literals to be a re-export of Qt functions. (#6175)
 
 ## 2.5.3